+++ movies-library/package.json
@@ -7,19 +7,17 @@
     "@types/react": "^18.2.22",
     "axios": "^1.5.1",
     "formik": "^2.4.5",
+    "next": "^14.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-portal": "^4.2.2",
-    "react-router-dom": "^6.16.0",
-    "react-scripts": "5.0.1",
     "styled-components": "^6.0.8",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test jest",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
