+++ movies-library/src/components/Commons/Icons/CloseIcon.js
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 
 const CloseIcon = (props) => (
