+++ movies-library/src/components/Commons/InputField/InputField.js
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import styled from "styled-components";
