+++ movies-library/src/components/Commons/Label/Label.js
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import styled from "styled-components";
