+++ movies-library/src/components/Counter/Counter.js
@@ -1,3 +1,4 @@
+"use client";
 import React, { useState } from "react";
 import Button from "../Commons/Button";
