+++ movies-library/src/components/FilterSection/GenreSelector.js
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 function GenreSelector(props) {
