+++ movies-library/src/components/FilterSection/SortControl.js
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 function SortControl(props) {
