+++ movies-library/src/components/Movie/MovieDetails.js
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 function MovieDetails(props) {
