+++ movies-library/src/components/Movie/MovieDetailsCheck.js
@@ -1,14 +1,18 @@
-import { Navigate, useParams } from "react-router-dom";
+"use client";
+
 import { useMemo } from "react";
 import React from "react";
 import MovieDetails from "./MovieDetails";
-import Dialog from "components/Dialog/Dialog";
-import AddMovieForm from "components/MovieForm/AddMovieForm/AddMovieForm";
-import { FormModel } from "components/MovieForm/FormModel";
+import Dialog from "../Dialog/Dialog";
+import AddMovieForm from "..//MovieForm/AddMovieForm/AddMovieForm";
+import { FormModel } from "..//MovieForm/FormModel";
+import { useRouter } from "next/router";
+import Navigate from "../../lib/Navigate";
 
 const checkMovieDetails =
   (movies, navigateMainPage, checkType, handleMovieEdit) => () => {
-    const { movieId } = useParams();
+    const router = useRouter();
+    const { movieId } = router.query;
 
     const movieDetails = useMemo(
       () => (movieId ? movies.find((movie) => movie.title === movieId) : null),
@@ -31,7 +35,7 @@ const checkMovieDetails =
         />
       )
     ) : (
-      <Navigate to="/" />
+      <Navigate to="/" replace />
     );
   };
