+++ movies-library/src/components/Movie/MoviesList.js
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import MovieTile from "./MovieTile";
