+++ movies-library/src/components/MovieForm/AddMovieForm/AddMovieButton.js
@@ -1,17 +1,13 @@
-import Button from "components/Commons/Button";
-import { useNavigate } from "react-router-dom";
+"use client";
+import Button from "../../../components/Commons/Button";
 import React from "react";
+import Link from "next/link";
 
 function AddMovieButton() {
-  const navigate = useNavigate();
-  const handleButtonClick = () => navigate("/new");
-
   return (
-    <Button
-      className="m-search__button--add"
-      title={"+ ADD MOVIE"}
-      clickAction={handleButtonClick}
-    />
+    <Link href="/new">
+      <Button className="m-search__button--add" title={"+ ADD MOVIE"} />
+    </Link>
   );
 }
 export default AddMovieButton;
