+++ movies-library/src/components/MovieForm/AddMovieForm/EditMovieButton.js
@@ -1,11 +1,12 @@
-import Button from "components/Commons/Button";
-import { useNavigate } from "react-router-dom";
+"use client";
+import Button from "../../Commons/Button";
+// import { useNavigate } from "react-router-dom";
 import React from "react";
 
 function EditMovieButton(props) {
-  const navigate = useNavigate();
+  //   const navigate = useNavigate();
   const handleButtonClick = () => {
-    navigate(`${props.title}/edit`);
+    // navigate(`${props.title}/edit`);
   };
 
   return <Button title={"EDIT MOVIE"} clickAction={handleButtonClick} />;
