+++ movies-library/src/components/MovieForm/DeleteMovieForm/DeleteMovieForm.js
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import styled from "styled-components";
