+++ movies-library/src/components/MovieForm/GenreSelector/GenreCheckBox/GenreCheckBox.js
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import styled from "styled-components";
