+++ movies-library/src/components/MovieForm/GenreSelector/GenreSelectorInput/GenreSelectorInput.js
@@ -1,3 +1,4 @@
+"use client";
 import { useState } from "react";
 import React from "react";
 import styled from "styled-components";
