+++ movies-library/src/components/MovieForm/SuccessInfo/SuccessInfo.js
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import styled from "styled-components";
 import SuccessIcon from "../../Commons/Icons/SuccessTick";
