+++ movies-library/src/components/Search/SearchSection.stories.js
@@ -1,13 +1,13 @@
-/* eslint-disable import/no-anonymous-default-export */
-import React from "react";
-import SearchForm from "./SearchForm";
+// /* eslint-disable import/no-anonymous-default-export */
+// import React from "react";
+// import SearchForm from "./SearchForm";
 
-export default {
-  title: "Example/Search",
-  component: SearchForm,
-};
+// export default {
+//   title: "Example/Search",
+//   component: SearchForm,
+// };
 
-function Story(args) {
-  return <SearchForm {...args} />;
-}
-export const Default2 = Story.bind({});
+// function Story(args) {
+//   return <SearchForm {...args} />;
+// }
+// export const Default2 = Story.bind({});
