+++ movies-library/src/components/root.js
@@ -1,10 +1,8 @@
-import { Outlet } from "react-router-dom";
+// import { Outlet } from "react-router-dom";
 import React from "react";
 
 const Root = ({ className }) => (
-  <div className={className}>
-    <Outlet />
-  </div>
+  <div className={className}>{/* <Outlet /> */}</div>
 );
 
 export default Root;
